Additional grid search refactoring
