+++ bluecast/ml_modelling/base_classes.py
@@ -4,7 +4,7 @@ import logging
 import warnings
 from abc import ABC, abstractmethod
 from datetime import datetime
-from typing import Any, List, Literal, Optional, Tuple, TypeVar, Union
+from typing import Any, Dict, List, Literal, Optional, Tuple, TypeVar, Union
 
 import numpy as np
 import optuna
@@ -258,6 +258,43 @@ class XgboostBaseModel:
         y_test: pd.Series,
     ):
         raise NotImplementedError("Method fine_tune has not been defined.")
+
+    def create_fine_tune_search_space(self) -> Dict[str, np.array]:
+        if (
+            isinstance(self.conf_params_xgboost.params["min_child_weight"], float)
+            and isinstance(self.conf_params_xgboost.params["lambda"], float)
+            and isinstance(self.conf_params_xgboost.params["gamma"], float)
+            and isinstance(self.conf_params_xgboost.params["eta"], float)
+        ):
+            search_space = {
+                "min_child_weight": np.linspace(
+                    self.conf_params_xgboost.params["min_child_weight"] * 0.9,
+                    self.conf_params_xgboost.params["min_child_weight"] * 1.1,
+                    self.conf_training.gridsearch_nb_parameters_per_grid,
+                    dtype=float,
+                ),
+                "lambda": np.linspace(
+                    self.conf_params_xgboost.params["lambda"] * 0.9,
+                    self.conf_params_xgboost.params["lambda"] * 1.1,
+                    self.conf_training.gridsearch_nb_parameters_per_grid,
+                    dtype=float,
+                ),
+                "gamma": np.linspace(
+                    self.conf_params_xgboost.params["gamma"] * 0.9,
+                    self.conf_params_xgboost.params["gamma"] * 1.1,
+                    self.conf_training.gridsearch_nb_parameters_per_grid,
+                    dtype=float,
+                ),
+                "eta": np.linspace(
+                    self.conf_params_xgboost.params["eta"] * 0.9,
+                    self.conf_params_xgboost.params["eta"] * 1.1,
+                    self.conf_training.gridsearch_nb_parameters_per_grid,
+                    dtype=float,
+                ),
+            }
+            return search_space
+        else:
+            raise ValueError("Some parameters are not floats or strings")
 
     def orchestrate_hyperparameter_tuning(
         self,
