+++ bluecast/ml_modelling/xgboost_regression.py
@@ -595,7 +595,7 @@ class XgboostModelRegression(XgboostBaseModel):
                     y_test,
                 )
             else:
-                # make regression cv startegy stratified
+                # make regression cv strategy stratified
                 le = LabelEncoder()
                 y_binned = le.fit_transform(pd.qcut(y_train, 10, duplicates="drop"))
                 skf = RepeatedStratifiedKFold(
@@ -647,32 +647,7 @@ class XgboostModelRegression(XgboostBaseModel):
             and isinstance(self.conf_params_xgboost.params["gamma"], float)
             and isinstance(self.conf_params_xgboost.params["eta"], float)
         ):
-            search_space = {
-                "min_child_weight": np.linspace(
-                    self.conf_params_xgboost.params["min_child_weight"] * 0.9,
-                    self.conf_params_xgboost.params["min_child_weight"] * 1.1,
-                    self.conf_training.gridsearch_nb_parameters_per_grid,
-                    dtype=float,
-                ),
-                "lambda": np.linspace(
-                    self.conf_params_xgboost.params["lambda"] * 0.9,
-                    self.conf_params_xgboost.params["lambda"] * 1.1,
-                    self.conf_training.gridsearch_nb_parameters_per_grid,
-                    dtype=float,
-                ),
-                "gamma": np.linspace(
-                    self.conf_params_xgboost.params["gamma"] * 0.9,
-                    self.conf_params_xgboost.params["gamma"] * 1.1,
-                    self.conf_training.gridsearch_nb_parameters_per_grid,
-                    dtype=float,
-                ),
-                "eta": np.linspace(
-                    self.conf_params_xgboost.params["eta"] * 0.9,
-                    self.conf_params_xgboost.params["eta"] * 1.1,
-                    self.conf_training.gridsearch_nb_parameters_per_grid,
-                    dtype=float,
-                ),
-            }
+            search_space = self.create_fine_tune_search_space()
         else:
             ValueError("Some parameters are not floats or strings")
